added ShaderProgram class - basic
Created a basic ShaderProgram class that will read different shader files (should be compatible with any type of shader files accepted by openGL), compile the shaders, link and attach them into a program.

Ideally, everything should be done with the call of the constructor which would generate a program id (the only public member of the class).